// web/src/api/generated/model/recurringDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { RecurrenceFrequency } from './recurrenceFrequency';
import type { RecurringTemplate } from './recurringTemplate';

export interface RecurringDto {
  id: string;
  name: string;
  template: RecurringTemplate;
  templateKind: string;
  frequency: RecurrenceFrequency;
  nextRunDate: string;
  /** @nullable */
  endDate: string | null;
  autoIssue: boolean;
  isActive: boolean;
  /** @nullable */
  lastRunDate: string | null;
  runCount: number;
}